+++ pychron/experiment/automated_run/data_collector.py
@@ -351,18 +351,22 @@ class DataCollector(Consoleable):
                 if not action_conditional.resume:
                     return 'break'
     @property
+    def arar_age(self):
+        return self.automated_run.arar_age
+
+    @property
     def plot_panel(self):
         return self.automated_run.plot_panel
 
     @property
     def truncation_conditionals(self):
-        return self.automated_run.truncation_conditions
+        return self.automated_run.truncation_conditionals
 
     @property
     def termination_conditionals(self):
-        return self.automated_run.termination_conditions
+        return self.automated_run.termination_conditionals
 
     @property
     def action_conditionals(self):
-        return self.automated_run.action_conditions
+        return self.automated_run.action_conditionals
 #============= EOF =============================================